Move filterConfigs to core
Fixes #3056
