+++ src/core/utils.js
@@ -577,3 +577,15 @@ export const buildFormData = (data) => {
   }
   return formArr.join("&")
 }
+
+export const filterConfigs = (configs, allowed) => {
+    let i, filteredConfigs = {}
+
+    for (i in configs) {
+        if (allowed.indexOf(i) !== -1) {
+            filteredConfigs[i] = configs[i]
+        }
+    }
+
+    return filteredConfigs
+}
